AidSense/requirements.txt
torch>=2.0.0
transformers>=4.30.0
diffusers>=0.20.0
streamlit>=1.24.0
numpy>=1.24.0
pandas>=2.0.0
requests>=2.28.0
opencv-python>=4.7.0
scikit-learn>=1.2.2
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.1
python-dotenv>=1.0.0
langchain>=0.0.180
sentence-transformers>=2.2.2
